perf(limiter): target the exact refresh submultiple, not a rounded frame rate
GetTargetFrameRate rounded refresh/divisor to a whole frame rate, which breaks the
one property that matters for pacing: the target is meant to be a submultiple of
the display refresh. At 165 Hz a divisor of 4 became 41 fps (24.390 ms) instead of
41.25 (24.242 ms, exactly four refresh intervals) -- a 0.61% error that drifts a
full refresh interval about once a second. The display then shows one long
interval followed by a short one, which is exactly what the capture contains:

  ... 24.8, 55.8, [9.8], 7.9, 23.8 ...
  ... 25.5, 123.8, [8.5], 8.3, 8.3, 7.6, 8.5 ...

The divisors that happen to divide exactly, 1 and 3, did not show it.

Return the exact value. Measured with a frozen camera, no frame generation:

  target    displayed        stddev            spikes >2x median
  165       165.0 -> 165.0   1.389 -> 0.579    0.96% -> 0.05%
  82.5       83.0 ->  82.5   6.580 -> 0.897    1.51% -> 0.00%
  41.25      41.0 ->  41.3  12.108 -> 1.156    1.43% -> 0.00%

Worst-case animation error falls from 119.6 ms to 12.4 ms at the 41 fps target.
Every target now lands within 0.1% instead of being rounded away from the refresh
submultiple it was derived from.

+++ src/Features/Upscaling.h
@@ -152,7 +152,7 @@ public:
 	/** @brief Returns the monitor refresh rate in hertz. */
 	[[nodiscard]] int GetMonitorRefreshRate() const;
 	/** @brief Returns the configured frame-rate cap, or zero when uncapped. */
-	[[nodiscard]] int GetTargetFrameRate() const;
+	[[nodiscard]] double GetTargetFrameRate() const;
 	/** @brief Returns the rendered-frame cap after accounting for fixed frame generation. */
 	[[nodiscard]] double GetRenderedFrameRateLimit() const;
 	/** @brief Returns the fixed DLSS-G multiplier clamped to the reported hardware limit. */
